flower changes for real edge
